뽀모도로페이지에 들어갈 screen 파일 분리
