+++ src/renderer/features/pomodoro/index.ts
@@ -1,1 +1,2 @@
 export * from './hooks/use-add-pomodoro';
+export * from './util/cat-name-map';
