+++ src/renderer/features/pomodoro/util/cat-name-map.ts
@@ -1,0 +1,8 @@
+import { CatType } from '@/entities/cat';
+
+export const catNameMap = (type: CatType) => {
+  if (type === 'CHEESE') return '치즈냥';
+  if (type === 'BLACK') return '까만냥';
+  if (type === 'THREE_COLOR') return '삼색냥';
+  return '';
+};
