+++ src/renderer/features/time/index.ts
@@ -1,2 +1,3 @@
 export * from './ui/chage-time-dialog';
 export * from './hooks/use-focus-notification';
+export * from './ui/time';
