+++ src/renderer/features/time/ui/time.tsx
@@ -1,0 +1,17 @@
+import { cn, padNumber } from '@/shared/utils';
+
+type TimeProps = {
+  minutes: number;
+  seconds: number;
+  className?: string;
+};
+
+export const Time = ({ minutes, seconds, className }: TimeProps) => {
+  return (
+    <div className={cn('flex items-center justify-center', className)}>
+      <span className="tabular-nums">{padNumber(minutes)}</span>
+      <span>:</span>
+      <span className="tabular-nums">{padNumber(seconds)}</span>
+    </div>
+  );
+};
