+++ src/renderer/widgets/pomodoro/ui/rest-screen.tsx
@@ -1,0 +1,3 @@
+export const RestScreen = () => {
+  return <div>RestScreen</div>;
+};
