+++ src/renderer/widgets/pomodoro/ui/rest-wait-screen.tsx
@@ -1,0 +1,3 @@
+export const RestWaitScreen = () => {
+  return <div>RestWaitScreen</div>;
+};
